feat: Add Sensors class to handle sensors
get_distance will get the distance with an ultrasonic sensor.
get_presence will use an IR proximity sensor to detect a presence, not
developped yet.

All of this will be used to detect if Maqueen is next to the trafic light
to stop it or not. Only one method will be used.